Move S2Processor to Processors and add filter to s2masking
