+++ dep_tools/processors.py
@@ -1,9 +1,10 @@
 from abc import ABC, abstractmethod
-from typing import Tuple
 
 from xarray import DataArray
 
-from .landsat_utils import mask_clouds
+from .landsat_utils import mask_clouds as mask_clouds_landsat
+from .s2_utils import harmonize_to_old
+from .s2_utils import mask_clouds as mask_clouds_s2
 from .utils import scale_and_offset
 
 
@@ -31,7 +32,7 @@ class LandsatProcessor(Processor):
 
     def process(self, xr: DataArray) -> DataArray:
         if self.mask_clouds:
-            xr = mask_clouds(xr, **self.mask_kwargs)
+            xr = mask_clouds_landsat(xr, **self.mask_kwargs)
         if self.scale_and_offset:
             # These values only work for SR bands of landsat. Ideally we could
             # read from metadata. _Really_ ideally we could just pass "scale"
@@ -39,5 +40,28 @@ class LandsatProcessor(Processor):
             scale = 0.0000275
             offset = -0.2
             xr = scale_and_offset(xr, scale=[scale], offset=offset)
+
+        return xr
+
+
+class S2Processor(Processor):
+    def __init__(
+        self,
+        send_area_to_processor: bool = False,
+        scale_and_offset: bool = True,
+        mask_clouds: bool = True,
+        mask_clouds_kwargs: dict = dict(),
+    ) -> None:
+        super().__init__(send_area_to_processor)
+        self.scale_and_offset = scale_and_offset
+        self.mask_clouds = mask_clouds
+        self.mask_kwargs = mask_clouds_kwargs
+
+    def process(self, xr: DataArray) -> DataArray:
+        if self.mask_clouds:
+            xr = mask_clouds_s2(xr, **self.mask_kwargs)
+
+        if self.scale_and_offset:
+            xr = harmonize_to_old(xr)
 
         return xr
